# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway для микросервисов Forklore",
        "title": "Forklore API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/rankings/monthly": {
            "get": {
                "description": "Возвращает таблицу лидеров по рейтингу за последние 30 дней",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rankings"
                ],
                "summary": "Получить месячный рейтинг",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RankingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rankings/overall": {
            "get": {
                "description": "Возвращает таблицу лидеров по общему рейтингу пользователей",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rankings"
                ],
                "summary": "Получить общий рейтинг",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RankingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rankings/subject/{tag_id}": {
            "get": {
                "description": "Возвращает таблицу лидеров по выбранному предметному тегу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rankings"
                ],
                "summary": "Получить предметный рейтинг",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID предметного тега",
                        "name": "tag_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RankingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "models.RankingEntryResponse": {
            "type": "object",
            "properties": {
                "active_months_count": {
                    "type": "integer"
                },
                "active_weeks_last_8": {
                    "type": "integer"
                },
                "activity_points_30d": {
                    "type": "integer"
                },
                "activity_points_total": {
                    "type": "integer"
                },
                "avatar_url": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "followers_count": {
                    "type": "integer"
                },
                "followers_gained_30d": {
                    "type": "integer"
                },
                "forks_received_30d": {
                    "type": "integer"
                },
                "forks_received_total": {
                    "type": "integer"
                },
                "public_repositories_count": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "stars_received_30d": {
                    "type": "integer"
                },
                "stars_received_total": {
                    "type": "integer"
                },
                "subject_score": {
                    "type": "number"
                },
                "tag_id": {
                    "type": "string"
                },
                "title_label": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.RankingResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RankingEntryResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите \"Bearer\" и через пробел JWT access токен.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}